backend/pyproject.toml
[tool.poetry]
name = "timbre-backend"
version = "0.1.0"
description = "Timbre - ML-powered music recommendation engine backend"
authors = ["Conner Groth"]
readme = "README.md"
packages = [{include = "app"}, {include = "ingestion"}, {include = "models"}, {include = "services"}, {include = "routes"}, {include = "utils"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.103.1"
uvicorn = ">=0.23.2"
python-multipart = ">=0.0.6"
httpx = ">=0.25.0"
aiofiles = ">=23.0.0"
playwright = ">=1.40.0"
beautifulsoup4 = ">=4.12.0"
lxml = ">=4.9.0"
slowapi = ">=0.1.8"
python-dotenv = ">=1.0.0"
pydantic = ">=2.4.2"
pydantic-settings = ">=2.0.0"
sqlalchemy = ">=2.0.0"
alembic = ">=1.12.0"
psycopg2-binary = ">=2.9.0"
redis = ">=5.0.0"
celery = ">=5.3.0"
python-dateutil = ">=2.8.0"
requests = ">=2.31.0"
spotipy = ">=2.22.1"
pylast = ">=5.2.0"
supabase = ">=2.4.2"
sentence-transformers = ">=2.2.2"
numpy = ">=1.24.0"
scikit-learn = ">=1.3.0"
pandas = ">=2.0.0"
polars = ">=0.20.0"
openai = ">=1.0.0"
typer = ">=0.9.0"
selectolax = ">=0.3.0"
cloudscraper = "^1.2.71"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
black = ">=23.0.0"
flake8 = ">=6.0.0"
ruff = ">=0.1.0"
mypy = ">=1.0.0"
respx = ">=0.20.0"

[tool.poetry.scripts]
run-pipeline = "scripts.run_pipeline:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]